fix(activerecord): follow the AR test helper move to support/
#5361 moved src/test-helpers/ to src/support/; the two new migration
constraint tests still imported canonical-schema from the old path.
